// backend/src/common/enums/workspace-action.enum.ts
export enum WorkspaceAction {
  CREATED = 'CREATED',
  UPDATED = 'UPDATED',
  DELETED = 'DELETED',
  MEMBER_JOINED = 'MEMBER_JOINED',
  MEMBER_REMOVED = 'MEMBER_REMOVED',
  MEMBER_ROLE_UPDATED = 'MEMBER_ROLE_UPDATED',
  COLLABORATOR_ASSIGNED = 'COLLABORATOR_ASSIGNED',
  COLLABORATOR_REMOVED = 'COLLABORATOR_REMOVED',
  SEND_INVITATION_TO_JOIN = 'SEND_INVITATION_TO_JOIN',
}